requirements.txt
flask
youtube-transcript-api
nltk
python-pptx
google-auth
google-auth-oauthlib
google-auth-httplib2
google-api-python-client